api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
upstash-redis==1.2.0
google-cloud-tasks==2.17.1
google-cloud-storage==2.19.0
httpx==0.28.1
